Started the UI implementation.
Added a form. Have some items and x's and o's that aren't currently resources. The overall implementation of the whole thing is sketchy at best and will need tweaking.

Not all picturebox's are currently updating only the top left one. The settings will need to be copied over and the images being used updated as resources.